working test and production databases, complete GameSetupHelper

+++ composeApp/src/commonMain/kotlin/com/minutesock/dawordgame/core/domain/GameLanguage.kt
@@ -21,7 +21,8 @@ enum class GameLanguage {
         }
     }
 
-    val expectedWordSelectionCount: Int get() {
+    val expectedWordSelectionCount: Long
+        get() {
         return when (this) {
             English -> 2_315
             German -> 0
